IMPROVEMENT: add support of FusionCache
