+++ src/Webinex.Revoke.Abstractions/RevokeId.cs
@@ -1,15 +1,14 @@
-﻿namespace Webinex.Revoke
-{
-    public class RevokeId
-    {
-        public RevokeId(string kind, string value)
-        {
-            Kind = kind;
-            Value = value;
-        }
+﻿namespace Webinex.Revoke;
 
-        public string Kind { get; }
-        
-        public string Value { get; }
+public class RevokeId
+{
+    public RevokeId(string kind, string value)
+    {
+        Kind = kind;
+        Value = value;
     }
+
+    public string Kind { get; }
+        
+    public string Value { get; }
 }
