+++ src/Webinex.Revoke.All/Package.props
@@ -1,12 +1,13 @@
 ﻿<Project>
     <PropertyGroup>
         <Title>Webinex Revoke</Title>
-        <PackageVersion>0.0.2</PackageVersion>
+        <PackageVersion>1.0.0-rc1</PackageVersion>
         <Authors>Webinex Dev</Authors>
         <NoWarn>$(NoWarn);CS1591</NoWarn>
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-        <LangVersion>8.0</LangVersion>
-        <Nullable>disable</Nullable>
+        <LangVersion>12</LangVersion>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
         <Description>Easy to use black listing access token</Description>
         <Copyright>Webinex Dev</Copyright>
         <PackageProjectUrl>https://github.com/webinex/starter-kit</PackageProjectUrl>
