+++ src/Webinex.Revoke.All/Webinex.Revoke.All.csproj
@@ -2,13 +2,15 @@
     <Import Project="$(MSBuildThisFileDirectory)/Package.props" />
 
     <PropertyGroup>
-        <TargetFramework>netstandard2.1</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <PackageId>Webinex.Revoke.All</PackageId>
     </PropertyGroup>
 
     <ItemGroup>
       <ProjectReference Include="..\Webinex.Revoke.Abstractions\Webinex.Revoke.Abstractions.csproj" />
       <ProjectReference Include="..\Webinex.Revoke\Webinex.Revoke.csproj" />
+      <ProjectReference Include="..\Webinex.Revoke.FusionCache\Webinex.Revoke.FusionCache.csproj" />
+      <ProjectReference Include="..\Webinex.Revoke.DistributedCache\Webinex.Revoke.DistributedCache.csproj" />
       <ProjectReference Include="..\Webinex.Revoke.Middleware\Webinex.Revoke.Middleware.csproj" />
     </ItemGroup>
