+++ src/Webinex.Revoke.Middleware/ApplicationBuilderExtensions.cs
@@ -1,23 +1,21 @@
-﻿using System;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 
-namespace Webinex.Revoke.Middleware
+namespace Webinex.Revoke.Middleware;
+
+public static class ApplicationBuilderExtensions
 {
-    public static class ApplicationBuilderExtensions
+    /// <summary>
+    ///     Adds revoke middleware that checks incoming request for revoke.
+    ///     If request matches any revoke criteria. Request would be declined.
+    /// </summary>
+    /// <param name="app"><see cref="IApplicationBuilder"/></param>
+    /// <returns><see cref="IApplicationBuilder"/></returns>
+    public static IApplicationBuilder UseRevoke(
+        this IApplicationBuilder app)
     {
-        /// <summary>
-        ///     Adds revoke middleware that checks incoming request for revoke.
-        ///     If request matches any revoke criteria. Request would be declined.
-        /// </summary>
-        /// <param name="app"><see cref="IApplicationBuilder"/></param>
-        /// <returns><see cref="IApplicationBuilder"/></returns>
-        public static IApplicationBuilder UseRevoke(
-            this IApplicationBuilder app)
-        {
-            app = app ?? throw new ArgumentNullException(nameof(app));
+        app = app ?? throw new ArgumentNullException(nameof(app));
             
-            app.UseMiddleware<RevokeMiddleware>();
-            return app;
-        }
+        app.UseMiddleware<RevokeMiddleware>();
+        return app;
     }
 }
