+++ src/Webinex.Revoke.Middleware/DelegateRevokeAccessor.cs
@@ -1,22 +1,18 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
 
-namespace Webinex.Revoke.Middleware
+namespace Webinex.Revoke.Middleware;
+
+internal class DelegateRevokeAccessor : IRevokeAccessor
 {
-    internal class DelegateRevokeAccessor : IRevokeAccessor
+    private readonly Func<HttpContext, IEnumerable<RevokeId>> _accessor;
+
+    public DelegateRevokeAccessor(Func<HttpContext, IEnumerable<RevokeId>> accessor)
     {
-        private readonly Func<HttpContext, IEnumerable<RevokeId>> _accessor;
+        _accessor = accessor;
+    }
 
-        public DelegateRevokeAccessor(Func<HttpContext, IEnumerable<RevokeId>> accessor)
-        {
-            _accessor = accessor;
-        }
-
-        public Task<IEnumerable<RevokeId>> GetAsync(HttpContext context)
-        {
-            return Task.FromResult(_accessor(context));
-        }
+    public Task<IEnumerable<RevokeId>> GetAsync(HttpContext context)
+    {
+        return Task.FromResult(_accessor(context));
     }
 }
