+++ src/Webinex.Revoke.Middleware/ProtegoRevokeConfigurationExtensions.cs
@@ -1,21 +1,18 @@
-﻿using System;
+﻿namespace Webinex.Revoke.Middleware;
 
-namespace Webinex.Revoke.Middleware
+public static class ProtegoRevokeConfigurationExtensions
 {
-    public static class ProtegoRevokeConfigurationExtensions
+    public static IRevokeConfiguration AddMiddleware(
+        this IRevokeConfiguration revokeConfiguration,
+        Action<IRevokeMiddlewareConfiguration> configure)
     {
-        public static IRevokeConfiguration AddMiddleware(
-            this IRevokeConfiguration revokeConfiguration,
-            Action<IRevokeMiddlewareConfiguration> configure)
-        {
-            revokeConfiguration = revokeConfiguration ?? throw new ArgumentNullException(nameof(revokeConfiguration));
-            configure = configure ?? throw new ArgumentNullException(nameof(configure));
+        revokeConfiguration = revokeConfiguration ?? throw new ArgumentNullException(nameof(revokeConfiguration));
+        configure = configure ?? throw new ArgumentNullException(nameof(configure));
 
-            var configuration = new RevokeMiddlewareConfiguration(revokeConfiguration.Services);
-            configure(configuration);
-            configuration.Complete();
+        var configuration = new RevokeMiddlewareConfiguration(revokeConfiguration.Services);
+        configure(configuration);
+        configuration.Complete();
 
-            return revokeConfiguration;
-        }
+        return revokeConfiguration;
     }
 }
