+++ src/Webinex.Revoke.Middleware/RevokeClaim.cs
@@ -1,17 +1,14 @@
-﻿using System;
+﻿namespace Webinex.Revoke.Middleware;
 
-namespace Webinex.Revoke.Middleware
+internal class RevokeClaim
 {
-    internal class RevokeClaim
+    public RevokeClaim(string claimName, string revokeKind)
     {
-        public RevokeClaim(string claimName, string revokeKind)
-        {
-            ClaimName = claimName ?? throw new ArgumentNullException(nameof(claimName));
-            RevokeKind = revokeKind ?? throw new ArgumentNullException(nameof(revokeKind));
-        }
-
-        public string ClaimName { get; }
-        
-        public string RevokeKind { get; }
+        ClaimName = claimName ?? throw new ArgumentNullException(nameof(claimName));
+        RevokeKind = revokeKind ?? throw new ArgumentNullException(nameof(revokeKind));
     }
+
+    public string ClaimName { get; }
+        
+    public string RevokeKind { get; }
 }
