+++ src/Webinex.Revoke.sln
@@ -3,13 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke", "Webinex.Revoke\Webinex.Revoke.csproj", "{AC21392F-A8E0-4B19-B24E-B9A0FB377FED}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke.DistributedCache", "Webinex.Revoke.DistributedCache\Webinex.Revoke.DistributedCache.csproj", "{AC21392F-A8E0-4B19-B24E-B9A0FB377FED}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke.Abstractions", "Webinex.Revoke.Abstractions\Webinex.Revoke.Abstractions.csproj", "{11C4C9D9-E405-487F-BD98-1E4CA119AD26}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke.All", "Webinex.Revoke.All\Webinex.Revoke.All.csproj", "{7D9037E4-4FA4-4A34-942F-CC69F58B4BA3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke.Middleware", "Webinex.Revoke.Middleware\Webinex.Revoke.Middleware.csproj", "{557CE670-7073-4398-B0FD-2D3F78817E9C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke.FusionCache", "Webinex.Revoke.FusionCache\Webinex.Revoke.FusionCache.csproj", "{42065D30-D368-4DDF-AE68-8C000499AEB6}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Webinex.Revoke", "Webinex.Revoke\Webinex.Revoke.csproj", "{81A30515-0735-4F6E-B319-3BA5ED3D3AEB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -36,5 +40,13 @@ Global
 		{557CE670-7073-4398-B0FD-2D3F78817E9C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{557CE670-7073-4398-B0FD-2D3F78817E9C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{557CE670-7073-4398-B0FD-2D3F78817E9C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42065D30-D368-4DDF-AE68-8C000499AEB6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42065D30-D368-4DDF-AE68-8C000499AEB6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42065D30-D368-4DDF-AE68-8C000499AEB6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42065D30-D368-4DDF-AE68-8C000499AEB6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{81A30515-0735-4F6E-B319-3BA5ED3D3AEB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{81A30515-0735-4F6E-B319-3BA5ED3D3AEB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{81A30515-0735-4F6E-B319-3BA5ED3D3AEB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{81A30515-0735-4F6E-B319-3BA5ED3D3AEB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
