+++ src/Webinex.Revoke/Revoke.cs
@@ -1,29 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Webinex.Revoke.Stores;
+﻿namespace Webinex.Revoke;
 
-namespace Webinex.Revoke
+internal class Revoke : IRevoke
 {
-    internal class Revoke : IRevoke
+    private readonly IRevokeWriteStore _revokeWriteStore;
+    private readonly IRevokeReadStore _revokeReadStore;
+
+    public Revoke(IRevokeWriteStore revokeWriteStore, IRevokeReadStore revokeReadStore)
     {
-        private readonly IRevokeWriteStore _revokeWriteStore;
-        private readonly IRevokeReadStore _revokeReadStore;
+        _revokeWriteStore = revokeWriteStore;
+        _revokeReadStore = revokeReadStore;
+    }
 
-        public Revoke(IRevokeWriteStore revokeWriteStore, IRevokeReadStore revokeReadStore)
-        {
-            _revokeWriteStore = revokeWriteStore;
-            _revokeReadStore = revokeReadStore;
-        }
+    public async Task<bool> RevokedAnyAsync(IEnumerable<RevokeId> ids, DateTime issued)
+    {
+        return await _revokeReadStore.RevokedAnyAsync(ids, issued);
+    }
 
-        public async Task<bool> RevokedAnyAsync(IEnumerable<RevokeId> ids, DateTime issued)
-        {
-            return await _revokeReadStore.RevokedAnyAsync(ids, issued);
-        }
-
-        public async Task RevokeAsync(IEnumerable<RevokeArgs> revokes)
-        {
-            await _revokeWriteStore.RevokeAsync(revokes);
-        }
+    public async Task RevokeAsync(IEnumerable<RevokeArgs> revokes)
+    {
+        await _revokeWriteStore.RevokeAsync(revokes);
     }
 }
