--- src/Webinex.Revoke/Stores.DistributedCache/IDistributedCacheSettings.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace Webinex.Revoke.Stores.DistributedCache
-{
-    internal interface IDistributedCacheSettings
-    {
-        string Prefix { get; }
-        string Separator { get; }
-        Func<RevokeId, TimeSpan> Expiration { get; }
-        string ClientCachePrefix { get; }
-        TimeSpan ClientCacheExpiration { get; }
-    }
-}
